Working solution with new IR sensor
PID regulation.
Plots added.
Cleaned up code.
Comments added.

+++ Project/Project/src/GlobalVariables.h
@@ -2,14 +2,15 @@
  * GlobalVariables.h
  *
  * Created: 11/29/2016 13:29:01
- *  Author: Nadia
+ *  Author: Nadia, Elvin
  */ 
 
 
 #ifndef GLOBALVARIABLES_H_
 #define GLOBALVARIABLES_H_
 
-extern uint8_t setpoint;
+/* Hold variables that is sent between Arduino and MATLAB as well as between PIDControl.c and MatlabPlot.c */
+extern uint16_t setpoint;
 extern double dT;
 extern double Kp;
 extern double Ti;
